Add autorun mode
This allows a migration to be marked (via a filename prefix) such that if the file changes
it will be re-run automatically without requiring --force. The use case is for stored
procedures where it's handier to have one file that changes when the proc is updated
rather than recreating it with several migrations.